Fixed "Forest map" not showing bug
This is a strange and special case.
The original name given for the map is "Forest  (easy mode)". But the website isn't showing it. Because, the 'if' condition in the double 'for' loops inside the 'alphaSort' function is not getting satisfied just for this map. Turns out, it's all just due to the double space inside the map's original name.
HTML renders double spaces as single space. So the 'innerText' property will be with only a single space.
In this function, I was comparing 'innerHTML' with 'innerText', which was resulting false as there are more than one space.
So yeah.. just changed that.
Pretty unique bug which is hard to find, but requires a really small fix.