Added Search Feature to the Website
Crappiest code I've ever written.
Goodluck future me, in figuring out the logic I've just written for cross-implementation of search feature.
Known Bug: When the maps are filtered, sorted reversely and a keyword is searched and then sorted back again alphabetically, the search is applied to the whole maps, but not just the filtered maps.